Project directory changeS
