+++ src/pages/quotegenerator/quotegenerator.tsx
@@ -31,10 +31,9 @@ import {
 import './profile.css';
 
 //Dynamic data reference
-import rizzCard from '../assets/json/rizzCard.json';
+import rizzCard from '../../assets/json/rizzCard.json';
 
-const Profile: React.FC = () => {
-
+const QuoteGenerator: React.FC = () => {
 
   const [showAlert, setShowAlert] = useState(false);
   const [randomIndex, setRandomIndex] =  useState<number | null>(null); // State to store random index
@@ -65,47 +64,28 @@ const Profile: React.FC = () => {
     setRandomIndex(0); // Reset the index to 0
     setShowAlert(false); // Hide the alert
   };
-
-  
   
   return (
     <IonPage>
       <IonHeader>
         <IonToolbar>
-          <IonTitle>Profile</IonTitle>
+          <IonTitle>Quote Generator</IonTitle>
         </IonToolbar>
       </IonHeader>
       <IonContent fullscreen>
         <IonHeader collapse="condense">
           <IonToolbar>
-            <IonTitle size="large">Profile</IonTitle>
+            <IonTitle size="large">Quote Generator</IonTitle>
           </IonToolbar>
         </IonHeader>
 
         <IonCard>
-          {/*Avatar and name*/}
-          <IonGrid>
-            <IonRow>
-                <IonCol size="1.5" push="">
-                  <IonAvatar>
-                    <img alt="Silhouette of a person's head" src="../src/assets/img/avatar.png" />
-                  </IonAvatar>
-                </IonCol>
-                <IonCol size="9" pull="">
-                  <IonCardTitle>Cliff Amadeus F. Evangelio</IonCardTitle>
-                  <IonCardSubtitle>Instructor 1 - NBSC BSIT</IonCardSubtitle>
-                </IonCol>
-            </IonRow>
-          </IonGrid>
 
-          {/*Button Grid*/}
+          {/*Button Trigger*/}
           <IonGrid>
             <IonRow>
                 <IonCol size="" push="">
-
-                <IonButton id="present-alert" color="warning" expand="full" onClick={handleOpenAlert}>
-                  Quick Facts
-                </IonButton>
+                <IonButton id="present-alert" color="warning" expand="full" onClick={handleOpenAlert}>Click me</IonButton> 
                 <IonAlert
                   isOpen={showAlert}
                   onDidDismiss={handleAlertDismiss} // Call the handleAlertDismiss function when the alert is closed
@@ -114,46 +94,13 @@ const Profile: React.FC = () => {
                   message={renderRandomMessage()}
                   buttons={['Close']}
                 />
-
-                </IonCol>
-                <IonCol size="" pull="">
-                  <IonButton fill="clear" id="open-action-sheet" expand="block">Open Action Sheet</IonButton>
                 </IonCol>
             </IonRow>
           </IonGrid>
-          <IonCardContent>Quote: Nothing more, nothing less.</IonCardContent>
-             
-          {/* Actionsheet*/}
-          <IonActionSheet
-            trigger="open-action-sheet"
-            header="Actions"
-            buttons={[
-              {
-                text: 'Delete',
-                role: 'destructive',
-                data: {
-                  action: 'delete',
-                },
-              },
-              {
-                text: 'Share',
-                data: {
-                  action: 'share',
-                },
-              },
-              {
-                text: 'Cancel',
-                role: 'cancel',
-                data: {
-                  action: 'cancel',
-                },
-              },
-            ]}
-          ></IonActionSheet>
         </IonCard>  
       </IonContent>
     </IonPage>
   );
 };
 
-export default Profile;
+export default  QuoteGenerator;
